*added refraction *added new scenes *fixed sphere intersection *fixed shadow ray computing *added checking for being inside of an object *made trace & shade functions smaller
