+++ phantasm/scenes/glass-ball.json
@@ -1,0 +1,52 @@
+{
+   "depth" : 10,
+
+   "background" : [255, 0, 255],
+
+   "camera": 
+       {
+        "position" : [0, 0, -25],
+        "lookAt"   : [0, 0, 0],
+        "up"       : [0, 1, 0],
+        "viewAngle": 60
+       },
+
+   "lights": 
+       [
+          {
+           "type"     : "point",
+           "position" : [10, -10, -10],
+           "color"    : [255, 255, 255]
+          },
+
+          {
+           "type"     : "point",
+           "position" : [-10, -10, -10],
+           "color"    : [255, 255, 255]
+          }
+
+       ],
+
+   "objects": 
+      [                        
+       
+       {
+           "type" :    "sphere",
+           "position": [9.0, 0.0, -4.0],
+           "material": 
+                       {
+                        "Ka"    : 0.3, 
+                        "Ks"    : 0.4, 
+                        "Kd"    : 0.5, 
+                        "Kr"    : 0.2, 
+                        "Kt"    : 0.6,
+                        "Phong" : 10.0, 
+                        "Refl"  : false, 
+                        "Refr"  : 1.5
+                       },
+           "color":    [100, 100, 100], 
+           "radius":   5.0
+       }
+     ]
+}
+
