+++ phantasm/scenes/two-mirrors.json
@@ -5,7 +5,7 @@
 
    "camera": 
        {
-        "position" : [0, 0, -20],
+        "position" : [0, 0, -25],
         "lookAt"   : [0, 0, 0],
         "up"       : [0, 1, 0],
         "viewAngle": 60
@@ -16,6 +16,12 @@
           {
            "type"     : "point",
            "position" : [10, -10, -10],
+           "color"    : [255, 255, 255]
+          },
+
+          {
+           "type"     : "point",
+           "position" : [-10, -10, -10],
            "color"    : [255, 255, 255]
           }
 
@@ -30,7 +36,9 @@
                        {
                         "Ka"    : 0.3, 
                         "Ks"    : 0.4, 
-                        "Kd"    : 0.5, 
+                        "Kd"    : 0.5,   
+                        "Kr"    : 0.0, 
+                        "Kt"    : 0.0,
                         "Phong" : 10.0, 
                         "Refl"  : false, 
                         "Refr"  : 0.0
@@ -46,7 +54,9 @@
                        {
                         "Ka"    : 0.3, 
                         "Ks"    : 0.4, 
-                        "Kd"    : 0.5, 
+                        "Kd"    : 0.5,   
+                        "Kr"    : 0.0, 
+                        "Kt"    : 0.0,
                         "Phong" : 10.0, 
                         "Refl"  : false, 
                         "Refr"  : 0.0
@@ -60,17 +70,36 @@
            "position": [0.0, 10.0, 0.0],
            "material": 
                        {
-                        "Ka"    : 0.3, 
-                        "Ks"    : 0.4, 
-                        "Kd"    : 0.5, 
+                        "Ka"    : 0.1, 
+                        "Ks"    : 0.1, 
+                        "Kd"    : 0.1,  
+                        "Kr"    : 0.7, 
+                        "Kt"    : 0.0,
                         "Phong" : 10.0, 
                         "Refl"  : true, 
                         "Refr"  : 0.0
                        },
-           "color":    [255, 255, 255], 
+           "color":    [10, 10, 10], 
            "normal":   [0, -1, 0]
-       }
+       },
 
+       {
+           "type" :    "plane",
+           "position": [0.0, 0.0, 100.0],
+           "material": 
+                       {
+                        "Ka"    : 0.3, 
+                        "Ks"    : 0.4, 
+                        "Kd"    : 0.5,  
+                        "Kr"    : 0.7, 
+                        "Kt"    : 0.0,
+                        "Phong" : 10.0, 
+                        "Refl"  : true, 
+                        "Refr"  : 0.0
+                       },
+           "color":    [10, 10, 10], 
+           "normal":   [0, 0, -1]
+       }
 
      ]
 }
