+++ phantasm/scenes/two-planes.json
@@ -1,14 +1,14 @@
 {
-   "depth" : 3,
+   "depth" : 10,
 
    "background" : [255, 0, 255],
 
    "camera": 
        {
-        "position" : [0, 0, -20],
+        "position" : [0, 0, -25],
         "lookAt"   : [0, 0, 0],
         "up"       : [0, 1, 0],
-        "viewAngle": 90
+        "viewAngle": 60
        },
 
    "lights": 
@@ -16,6 +16,12 @@
           {
            "type"     : "point",
            "position" : [10, -10, -10],
+           "color"    : [255, 255, 255]
+          },
+
+          {
+           "type"     : "point",
+           "position" : [-10, -10, -10],
            "color"    : [255, 255, 255]
           }
 
@@ -25,17 +31,17 @@
       [                        
        {
            "type" :    "sphere",
-           "position": [4.0, 0.0, 30.0],
+           "position": [4.0, 0.0, 80.0],
            "material": 
                        {
                         "Ka"    : 0.3, 
                         "Ks"    : 0.4, 
-                        "Kd"    : 0.5,  
+                        "Kd"    : 0.5,
                         "Kr"    : 0.0, 
-                        "Kt"    : 0.0,
+                        "Kt"    : 0.0,  
                         "Phong" : 10.0, 
                         "Refl"  : false, 
-                        "Refr"  : 0.0
+                        "Refr"  : 3.3
                        },
            "color":    [0, 0, 255], 
            "radius":   5.0
@@ -48,7 +54,7 @@
                        {
                         "Ka"    : 0.3, 
                         "Ks"    : 0.4, 
-                        "Kd"    : 0.5,   
+                        "Kd"    : 0.5, 
                         "Kr"    : 0.0, 
                         "Kt"    : 0.0,
                         "Phong" : 10.0, 
@@ -57,7 +63,26 @@
                        },
            "color":    [0, 255, 0], 
            "radius":   5.0
+       },  
+       
+       {
+           "type" :    "sphere",
+           "position": [9.0, 0.0, -4.0],
+           "material": 
+                       {
+                        "Ka"    : 0.3, 
+                        "Ks"    : 0.4, 
+                        "Kd"    : 0.5, 
+                        "Kr"    : 0.2, 
+                        "Kt"    : 0.6,
+                        "Phong" : 10.0, 
+                        "Refl"  : true, 
+                        "Refr"  : 1.5
+                       },
+           "color":    [100, 100, 100], 
+           "radius":   5.0
        },
+
 
       {
            "type" :    "plane",
@@ -66,7 +91,25 @@
                        {
                         "Ka"    : 0.3, 
                         "Ks"    : 0.4, 
-                        "Kd"    : 0.5,   
+                        "Kd"    : 0.5, 
+                        "Kr"    : 0.0, 
+                        "Kt"    : 0.0,
+                        "Phong" : 10.0, 
+                        "Refl"  : false, 
+                        "Refr"  : 0.0
+                       },
+           "color":    [100, 100, 100], 
+           "normal":   [0, -1, 0]
+       },
+
+       {
+           "type" :    "plane",
+           "position": [0.0, 0.0, 100.0],
+           "material": 
+                       {
+                        "Ka"    : 0.3, 
+                        "Ks"    : 0.4, 
+                        "Kd"    : 0.5, 
                         "Kr"    : 0.0, 
                         "Kt"    : 0.0,
                         "Phong" : 10.0, 
@@ -74,9 +117,8 @@
                         "Refr"  : 0.0
                        },
            "color":    [255, 0, 0], 
-           "normal":   [0, -1, 0]
+           "normal":   [0, 0, -1]
        }
-
 
      ]
 }
